[Core/Util] Cleanup some warnings in Util.cpp and Util.h             Add in frand function for SFMT

+++ src/shared/Util.h
@@ -38,7 +38,7 @@ std::string TimeToTimestampStr(time_t t);
 inline uint32 secsToTimeBitFields(time_t secs)
 {
     tm* lt = localtime(&secs);
-    return (lt->tm_year - 100) << 24 | lt->tm_mon  << 20 | (lt->tm_mday - 1) << 14 | lt->tm_wday << 11 | lt->tm_hour << 6 | lt->tm_min;
+    return uint32((lt->tm_year - 100) << 24 | lt->tm_mon  << 20 | (lt->tm_mday - 1) << 14 | lt->tm_wday << 11 | lt->tm_hour << 6 | lt->tm_min);
 }
 
 // Return a random number in the range min..max; (max-min) must be smaller than 32768.
